Refactored MeasuredText to TextMetrics and moved `measure` function into it.
