+++ src/core/math/index.js
@@ -7,7 +7,6 @@ export { default as Point } from './Point';
 export { default as ObservablePoint } from './ObservablePoint';
 export { default as Matrix } from './Matrix';
 export { default as GroupD8 } from './GroupD8';
-export { default as MeasuredText } from './MeasuredText';
 
 export { default as Circle } from './shapes/Circle';
 export { default as Ellipse } from './shapes/Ellipse';
